// web-client/src/generated/moor-common/uu-obj-id.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from "flatbuffers";

export class UuObjId {
    bb: flatbuffers.ByteBuffer | null = null;
    bb_pos = 0;
    __init(i: number, bb: flatbuffers.ByteBuffer): UuObjId {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    static getRootAsUuObjId(bb: flatbuffers.ByteBuffer, obj?: UuObjId): UuObjId {
        return (obj || new UuObjId()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    static getSizePrefixedRootAsUuObjId(bb: flatbuffers.ByteBuffer, obj?: UuObjId): UuObjId {
        bb.setPosition(bb.position() + flatbuffers.SIZE_PREFIX_LENGTH);
        return (obj || new UuObjId()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    packedValue(): bigint {
        const offset = this.bb!.__offset(this.bb_pos, 4);
        return offset ? this.bb!.readUint64(this.bb_pos + offset) : BigInt("0");
    }

    static startUuObjId(builder: flatbuffers.Builder) {
        builder.startObject(1);
    }

    static addPackedValue(builder: flatbuffers.Builder, packedValue: bigint) {
        builder.addFieldInt64(0, packedValue, BigInt("0"));
    }

    static endUuObjId(builder: flatbuffers.Builder): flatbuffers.Offset {
        const offset = builder.endObject();
        return offset;
    }

    static createUuObjId(builder: flatbuffers.Builder, packedValue: bigint): flatbuffers.Offset {
        UuObjId.startUuObjId(builder);
        UuObjId.addPackedValue(builder, packedValue);
        return UuObjId.endUuObjId(builder);
    }
}