// web-client/src/generated/moor-var/var-anonymous.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from "flatbuffers";

import { Obj } from "../moor-common/obj.js";

export class VarAnonymous {
    bb: flatbuffers.ByteBuffer | null = null;
    bb_pos = 0;
    __init(i: number, bb: flatbuffers.ByteBuffer): VarAnonymous {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    static getRootAsVarAnonymous(bb: flatbuffers.ByteBuffer, obj?: VarAnonymous): VarAnonymous {
        return (obj || new VarAnonymous()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    static getSizePrefixedRootAsVarAnonymous(bb: flatbuffers.ByteBuffer, obj?: VarAnonymous): VarAnonymous {
        bb.setPosition(bb.position() + flatbuffers.SIZE_PREFIX_LENGTH);
        return (obj || new VarAnonymous()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    obj(obj?: Obj): Obj | null {
        const offset = this.bb!.__offset(this.bb_pos, 4);
        return offset ? (obj || new Obj()).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
    }

    static startVarAnonymous(builder: flatbuffers.Builder) {
        builder.startObject(1);
    }

    static addObj(builder: flatbuffers.Builder, objOffset: flatbuffers.Offset) {
        builder.addFieldOffset(0, objOffset, 0);
    }

    static endVarAnonymous(builder: flatbuffers.Builder): flatbuffers.Offset {
        const offset = builder.endObject();
        builder.requiredField(offset, 4); // obj
        return offset;
    }

    static createVarAnonymous(builder: flatbuffers.Builder, objOffset: flatbuffers.Offset): flatbuffers.Offset {
        VarAnonymous.startVarAnonymous(builder);
        VarAnonymous.addObj(builder, objOffset);
        return VarAnonymous.endVarAnonymous(builder);
    }
}