// web-client/src/generated/moor-var/var-int.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from "flatbuffers";

export class VarInt {
    bb: flatbuffers.ByteBuffer | null = null;
    bb_pos = 0;
    __init(i: number, bb: flatbuffers.ByteBuffer): VarInt {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    static getRootAsVarInt(bb: flatbuffers.ByteBuffer, obj?: VarInt): VarInt {
        return (obj || new VarInt()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    static getSizePrefixedRootAsVarInt(bb: flatbuffers.ByteBuffer, obj?: VarInt): VarInt {
        bb.setPosition(bb.position() + flatbuffers.SIZE_PREFIX_LENGTH);
        return (obj || new VarInt()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    value(): bigint {
        const offset = this.bb!.__offset(this.bb_pos, 4);
        return offset ? this.bb!.readInt64(this.bb_pos + offset) : BigInt("0");
    }

    static startVarInt(builder: flatbuffers.Builder) {
        builder.startObject(1);
    }

    static addValue(builder: flatbuffers.Builder, value: bigint) {
        builder.addFieldInt64(0, value, BigInt("0"));
    }

    static endVarInt(builder: flatbuffers.Builder): flatbuffers.Offset {
        const offset = builder.endObject();
        return offset;
    }

    static createVarInt(builder: flatbuffers.Builder, value: bigint): flatbuffers.Offset {
        VarInt.startVarInt(builder);
        VarInt.addValue(builder, value);
        return VarInt.endVarInt(builder);
    }
}