// web-client/src/generated/moor-var/var-obj.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from "flatbuffers";

import { Obj } from "../moor-common/obj";

export class VarObj {
    bb: flatbuffers.ByteBuffer | null = null;
    bb_pos = 0;
    __init(i: number, bb: flatbuffers.ByteBuffer): VarObj {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    static getRootAsVarObj(bb: flatbuffers.ByteBuffer, obj?: VarObj): VarObj {
        return (obj || new VarObj()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    static getSizePrefixedRootAsVarObj(bb: flatbuffers.ByteBuffer, obj?: VarObj): VarObj {
        bb.setPosition(bb.position() + flatbuffers.SIZE_PREFIX_LENGTH);
        return (obj || new VarObj()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    obj(obj?: Obj): Obj | null {
        const offset = this.bb!.__offset(this.bb_pos, 4);
        return offset ? (obj || new Obj()).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
    }

    static startVarObj(builder: flatbuffers.Builder) {
        builder.startObject(1);
    }

    static addObj(builder: flatbuffers.Builder, objOffset: flatbuffers.Offset) {
        builder.addFieldOffset(0, objOffset, 0);
    }

    static endVarObj(builder: flatbuffers.Builder): flatbuffers.Offset {
        const offset = builder.endObject();
        builder.requiredField(offset, 4); // obj
        return offset;
    }

    static createVarObj(builder: flatbuffers.Builder, objOffset: flatbuffers.Offset): flatbuffers.Offset {
        VarObj.startVarObj(builder);
        VarObj.addObj(builder, objOffset);
        return VarObj.endVarObj(builder);
    }
}