// web-client/src/generated/moor-var/var-str.ts
// automatically generated by the FlatBuffers compiler, do not modify

import * as flatbuffers from "flatbuffers";

export class VarStr {
    bb: flatbuffers.ByteBuffer | null = null;
    bb_pos = 0;
    __init(i: number, bb: flatbuffers.ByteBuffer): VarStr {
        this.bb_pos = i;
        this.bb = bb;
        return this;
    }

    static getRootAsVarStr(bb: flatbuffers.ByteBuffer, obj?: VarStr): VarStr {
        return (obj || new VarStr()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    static getSizePrefixedRootAsVarStr(bb: flatbuffers.ByteBuffer, obj?: VarStr): VarStr {
        bb.setPosition(bb.position() + flatbuffers.SIZE_PREFIX_LENGTH);
        return (obj || new VarStr()).__init(bb.readInt32(bb.position()) + bb.position(), bb);
    }

    value(): string | null;
    value(optionalEncoding: flatbuffers.Encoding): string | Uint8Array | null;
    value(optionalEncoding?: any): string | Uint8Array | null {
        const offset = this.bb!.__offset(this.bb_pos, 4);
        return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
    }

    static startVarStr(builder: flatbuffers.Builder) {
        builder.startObject(1);
    }

    static addValue(builder: flatbuffers.Builder, valueOffset: flatbuffers.Offset) {
        builder.addFieldOffset(0, valueOffset, 0);
    }

    static endVarStr(builder: flatbuffers.Builder): flatbuffers.Offset {
        const offset = builder.endObject();
        builder.requiredField(offset, 4); // value
        return offset;
    }

    static createVarStr(builder: flatbuffers.Builder, valueOffset: flatbuffers.Offset): flatbuffers.Offset {
        VarStr.startVarStr(builder);
        VarStr.addValue(builder, valueOffset);
        return VarStr.endVarStr(builder);
    }
}