// web-client/src/generated/moor-var/var-union.ts
export enum VarUnion {
    NONE = 0,
    VarNone = 1,
    VarBool = 2,
    VarInt = 3,
    VarFloat = 4,
    VarStr = 5,
    VarObj = 6,
    VarErr = 7,
    VarList = 8,
    VarMap = 9,
    VarSym = 10,
    VarBinary = 11,
    VarFlyweight = 12,
    VarLambda = 13,
    VarAnonymous = 14,
}